# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/places": {
      "post": {
        "operationId": "PlacesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a house / office / locker",
        "tags": [
          "places"
        ]
      },
      "get": {
        "operationId": "PlacesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceListItemResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Places you own or that are shared with you",
        "tags": [
          "places"
        ]
      }
    },
    "/api/places/{placeId}": {
      "get": {
        "operationId": "PlacesController_findOne",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One place with its members and counts",
        "tags": [
          "places"
        ]
      },
      "patch": {
        "operationId": "PlacesController_update",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update place details",
        "tags": [
          "places"
        ]
      },
      "delete": {
        "operationId": "PlacesController_remove",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a place and everything inside it",
        "tags": [
          "places"
        ]
      }
    },
    "/api/places/{placeId}/activity": {
      "get": {
        "operationId": "PlacesController_activity",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityLogResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Who did what in this place",
        "tags": [
          "places"
        ]
      }
    },
...
    "/api/storages/by-label/{labelCode}": {
      "get": {
        "operationId": "StoragesController_findByLabel",
        "parameters": [
          {
            "name": "labelCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Look up a storage by its QR sticker code, across all your places",
        "tags": [
          "storages"
        ]
      }
    },
    "/api/places/{placeId}/storages": {
      "post": {
        "operationId": "StoragesController_create",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStorageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a storage. Pass parentId to nest it inside another one.",
        "tags": [
          "storages"
        ]
      },
      "get": {
        "operationId": "StoragesController_findAll",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentId",
            "required": false,
            "in": "query",
            "description": "Only direct children of this storage",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ROOM",
                "ALMIRAH",
                "WARDROBE",
                "CABINET",
                "SHELF",
                "DRAWER",
                "BED",
                "BOX",
                "SUITCASE",
                "BAG",
                "FRIDGE",
                "LOFT",
                "RACK",
                "DESK",
                "SAFE",
                "OTHER"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Match part of the storage name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rootOnly",
            "required": false,
            "in": "query",
            "description": "Return only root storages",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StorageListItemResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Flat list of storages — this feeds the dropdown when adding a thing",
        "tags": [
          "storages"
        ]
      }
    },
    "/api/places/{placeId}/storages/tree": {
      "get": {
        "operationId": "StoragesController_tree",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StorageNodeResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Nested tree of every storage in the place",
        "tags": [
          "storages"
        ]
      }
    },
    "/api/places/{placeId}/storages/{storageId}": {
      "get": {
        "operationId": "StoragesController_findOne",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One storage with its children and the items inside it",
        "tags": [
          "storages"
        ]
      },
      "patch": {
        "operationId": "StoragesController_update",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStorageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename, edit, or move a storage (with its whole subtree)",
        "tags": [
          "storages"
        ]
      },
      "delete": {
        "operationId": "StoragesController_remove",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageDeletedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a storage and its subtree; items inside become unassigned",
        "tags": [
          "storages"
        ]
      }
    },
...
  },
  "info": {
    "title": "FindMyStuff API",
    "description": "Remember where you kept your stuff.\n\nHierarchy: **Place** (home/office/locker) → **Storage** tree (room → almirah → shelf → box) → **Item**.\n\nEvery place-scoped route is nested under `/places/{placeId}` and is checked against your\nmembership of that place. Asking for a place you are not a member of returns **404**, not 403 —\nthe API never confirms that someone else's data exists.\n\nAll routes except signup, login, refresh, logout and health require\n`Authorization: Bearer <accessToken>`.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Home — Demo deleted"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "UserCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Shristi"
          },
          "email": {
            "type": "string",
            "example": "shristi@example.com"
          },
          "avatarMediaId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "CreatePlaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Home — Sector 62"
          },
          "type": {
            "type": "string",
            "enum": [
              "HOME",
              "OFFICE",
              "LOCKER",
              "VEHICLE",
              "STORAGE_UNIT",
              "OTHER"
            ],
            "default": "HOME"
          },
          "description": {
            "type": "string"
          },
          "widthValue": {
            "type": "number",
            "description": "Optional — most people never measure their rooms"
          },
          "lengthValue": {
            "type": "number"
          },
          "heightValue": {
            "type": "number"
          },
          "lengthUnit": {
            "type": "string",
            "enum": [
              "CM",
              "INCH"
            ]
          },
          "addressLine": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "example": 28.6139
          },
          "longitude": {
            "type": "number",
            "example": 77.209
          },
          "coverImageBase64": {
            "type": "string",
            "description": "Base64 cover photo of the place"
          }
        },
        "required": [
          "name"
        ]
      },
      "PlaceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Home — Sector 62"
          },
          "type": {
            "type": "string",
            "enum": [
              "HOME",
              "OFFICE",
              "LOCKER",
              "VEHICLE",
              "STORAGE_UNIT",
              "OTHER"
            ]
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "city": {
            "type": "object",
            "nullable": true,
            "example": "Noida"
          },
          "country": {
            "type": "object",
            "nullable": true,
            "example": "India"
          },
          "latitude": {
            "type": "object",
            "nullable": true,
            "example": 28.6139
          },
          "longitude": {
            "type": "object",
            "nullable": true,
            "example": 77.209
          },
          "coverMediaId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "description",
          "ownerId",
          "city",
          "country",
          "latitude",
          "longitude",
          "coverMediaId",
          "createdAt"
        ]
      },
      "PlaceListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Home — Sector 62"
          },
          "type": {
            "type": "string",
            "enum": [
              "HOME",
              "OFFICE",
              "LOCKER",
              "VEHICLE",
              "STORAGE_UNIT",
              "OTHER"
            ]
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "city": {
            "type": "object",
            "nullable": true,
            "example": "Noida"
          },
          "country": {
            "type": "object",
            "nullable": true,
            "example": "India"
          },
          "latitude": {
            "type": "object",
            "nullable": true,
            "example": 28.6139
          },
          "longitude": {
            "type": "object",
            "nullable": true,
            "example": 77.209
          },
          "coverMediaId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "myRole": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ],
            "description": "Your role in this place"
          },
          "memberCount": {
            "type": "number",
            "example": 2
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserCard"
            }
          },
          "storageCount": {
            "type": "number",
            "example": 5
          },
          "itemCount": {
            "type": "number",
            "example": 12,
            "description": "Excludes other people's private items"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "description",
          "ownerId",
          "city",
          "country",
          "latitude",
          "longitude",
          "coverMediaId",
          "createdAt",
          "myRole",
          "memberCount",
          "members",
          "storageCount",
          "itemCount"
        ]
      },
...
      "PlaceDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Home — Sector 62"
          },
          "type": {
            "type": "string",
            "enum": [
              "HOME",
              "OFFICE",
              "LOCKER",
              "VEHICLE",
              "STORAGE_UNIT",
              "OTHER"
            ]
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "city": {
            "type": "object",
            "nullable": true,
            "example": "Noida"
          },
          "country": {
            "type": "object",
            "nullable": true,
            "example": "India"
          },
          "latitude": {
            "type": "object",
            "nullable": true,
            "example": 28.6139
          },
          "longitude": {
            "type": "object",
            "nullable": true,
            "example": 77.209
          },
          "coverMediaId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/UserCard"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaceMemberResponse"
            }
          },
          "myRole": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ]
          },
          "storageCount": {
            "type": "number",
            "example": 5
          },
          "itemCount": {
            "type": "number",
            "example": 12
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "description",
          "ownerId",
          "city",
          "country",
          "latitude",
          "longitude",
          "coverMediaId",
          "createdAt",
          "owner",
          "members",
          "myRole",
          "storageCount",
          "itemCount"
        ]
      },
      "UpdatePlaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Home — Sector 62"
          },
          "type": {
            "type": "string",
            "enum": [
              "HOME",
              "OFFICE",
              "LOCKER",
              "VEHICLE",
              "STORAGE_UNIT",
              "OTHER"
            ],
            "default": "HOME"
          },
          "description": {
            "type": "string"
          },
          "widthValue": {
            "type": "number",
            "description": "Optional — most people never measure their rooms"
          },
          "lengthValue": {
            "type": "number"
          },
          "heightValue": {
            "type": "number"
          },
          "lengthUnit": {
            "type": "string",
            "enum": [
              "CM",
              "INCH"
            ]
          },
          "addressLine": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "example": 28.6139
          },
          "longitude": {
            "type": "number",
            "example": 77.209
          },
          "coverImageBase64": {
            "type": "string",
            "description": "Base64 cover photo of the place"
          }
        }
      },
      "ActivityLogResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string",
            "example": "ITEM_MOVED"
          },
          "entityType": {
            "type": "string",
            "example": "Item"
          },
          "entityId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "summary": {
            "type": "string",
            "example": "Moved \"Passport\" to Bedroom › Almirah › Top shelf"
          },
          "actor": {
            "$ref": "#/components/schemas/UserCard"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action",
          "entityType",
          "entityId",
          "summary",
          "actor",
          "createdAt"
        ]
      },
...
      "CreateStorageDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Almirah"
          },
          "type": {
            "type": "string",
            "enum": [
              "ROOM",
              "ALMIRAH",
              "WARDROBE",
              "CABINET",
              "SHELF",
              "DRAWER",
              "BED",
              "BOX",
              "SUITCASE",
              "BAG",
              "FRIDGE",
              "LOFT",
              "RACK",
              "DESK",
              "SAFE",
              "OTHER"
            ],
            "default": "OTHER"
          },
          "parentId": {
            "type": "string",
            "description": "Parent storage id. Omit for a top-level storage such as a room."
          },
          "description": {
            "type": "string"
          },
          "widthValue": {
            "type": "number"
          },
          "lengthValue": {
            "type": "number"
          },
          "heightValue": {
            "type": "number"
          },
          "lengthUnit": {
            "type": "string",
            "enum": [
              "CM",
              "INCH"
            ]
          },
          "latitude": {
            "type": "number",
            "description": "Useful for a bank locker or a box kept elsewhere"
          },
          "longitude": {
            "type": "number"
          },
          "coverImageBase64": {
            "type": "string",
            "description": "Base64 photo of the storage, e.g. the open almirah"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}